feat: Build MVP CLI, write directly to stdout, enforce Colorizer contract
    API refactor: This commit unexports colorizeRune to enforce the use of
Colorizer.Colorize(Reader, Writer). Colorize manages the IO, and is the contact
point with catbow for a consumer.

    This commit defines an interface called Cleanupper. Move Cleanup() from
ColorStrategy to Cleanupper.

    This commit creates a minimal CLI in main.go. CLI will interact
with the public facing catbow API to configure and orchestrate
a Colorizer. Some things of note:

        - CLI calls and prints ColorStrategy.Cleanup() to stdout when
        the Colorizer exits on an error or EOF. This properly cleans up
        SGR attributes.

        - CLI includes a function that creates a mock reader, string
        is supplied by shelling out to generate.py. Enable with -gen
        flag

    Alongside this the bufio.ReaderWriter has been traded out in favor of
an unbuffered io.Writer. The buffered model is not performant because a flush
after every write was required to guarantee output was displayed.
As a result this commit had to utilize a different method for reading the runes
which is achieved through a bufio.Scanner with a bufio.RuneScanner Split function