fix(core): implement and stabilize rainbow algorithm
Tests are included to ensure the correctness of the
algorithm. DESIGN.md contains an implementation of
the algorithm if a reference is needed.

A simple CLI was also created with facilities
to configure spread, frequency, and seed as well as
generating text through the generate_text.py script.

This commit also does the following:
- fix \r\n causing offset to be incremented twice
- fix gen_text output being all on one line
- move python implementation of rainbow algorithm
to DESIGN.md
- clean up unused / commented out code
- improve comments for the parameters to the
rainbow algorithm
- update imports and go.mod to reflect new github username
- improve information hiding around the library boundary
by unexporting internal types